requirements.txt
fastapi==0.115.12
Pillow==11.1.0
pydantic==2.11.1
pytesseract==0.3.13
python-dotenv==1.1.0
supabase==2.15.0
uvicorn==0.34.0
